requirements.txt
astroid==1.6.2
autopep8==1.3.4
certifi==2018.1.18
future==0.16.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.0.0
pycodestyle==2.3.1
pylint==1.8.3
python-telegram-bot==10.0.1
six==1.11.0
wrapt==1.10.11
